refactor: (step2) LoginMember 의존성 제거
