+++ src/main/java/nextstep/auth/domain/AuthenticatedMember.java
@@ -1,0 +1,5 @@
+package nextstep.auth.domain;
+
+public interface AuthenticatedMember {
+	boolean checkPassword(String credentials);
+}
